Fixed a problem where *somatic.indel.vcf.gz.tbi was added twice.
From Christina:

Hi Brian, Matthias and Michael,
 
In the process of indexing the results from the DKFZ/EMBL pipeline, Linda and Junjun discovered a few issues.  First, to be clear, there are 3 versions of uploads: (1) Docker 1.0.4 or earlier has 2 uploads, (2) Docker 1.0.5 has 1 upload, (3) HPC version also has 1 upload.  Examples of each version are provided below.  I’ve also downloaded the files from GNOS, so the contents are also listed.  Since there is no problem with the EMBL part, I’m only showing the DKFZ files.
 
Issue 1: For all Docker versions, one file *.somatic.indel.vcf.gz.tbi was listed twice in the XML (attached screenshot).  For 1.0.4 or earlier, the XMLs will soon become obsolete when the results are fixed and re-uploaded.  So we can leave them as they are.  For 1.0.5, Brian will need to fix the upload code, and the metadata for results already uploaded.
 
Issue 2: For all Docker versions, *.germline.indel.vcf.gz.tbi is missing.  Matthias, are you ok with not having this tabix in the results?  (I hope so)  Brian, please fix this in the upload code of the Docker.
 
Issue 3: For the HPC version, this example that Linda picked is 1 of 6 tumours from the same donor.  Unfortunately, multi-tumour results from a donor must be in 1 upload.  Otherwise, we can’t index it properly.  Sorry, Michael & Matthias, please remove these multi-tumor uploads and re-upload them as 1 GNOS analysis.  Please also prefix the file names with the tumor aliquot id as you’ve done with the single-tumour uploads.
 
One question: I notice there’s no tar ball associated with germline.indel in all 3 versions.  Is this intended?
 
Thanks,
Christina